continuing with recipe for cauldron
a hardcoded recipe now does "work" but something with how the jewel powder slot is handled is not 100%, it cannot accept a tag while the ingredient field can. Do the two input slots both need to share the ingredient field????

Also the handling of the recipe is not expandable, only handling one input, one jewel powder, and one output.